requirements.txt
autopep8==1.6.0
bpyutils==0.4.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
flake8==4.0.1
google-api-core==2.7.2
google-api-python-client==2.44.0
google-auth==2.6.5
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.0
httplib2==0.20.4
idna==3.3
mccabe==0.7.0
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.8
requests==2.27.1
rsa==4.8
six==1.16.0
toml==0.10.2
upgrade-requirements==1.7.0
uritemplate==4.1.1
urllib3==1.26.9
